feat: support multiple tracks in a stream.
Summary:
Right now it cost me 12 days and more than 100 problems to develop this project, most codes were written bt ChatGPT-4 except those fix bugs. Consider of I'm not familiar with multimedia field, I think this toy project is actually exceeded my expectation.
But it can't help me with more complicate features, such as:
1.Support multi client request one stream.
2.Use epoll network model for RTSP communication.
To be honest, I'm not satisfied with the design of multiple tracks, because it's too heavy to handle video and audio in two separate SessionHandler, I think it should have a better solution or framework which can implement it with a better performance.

Feature reserved:
1.Support transport RTP in TCP.
2.Implement RTCP.